+++ src/Ants/WallAnt.cpp
@@ -5,4 +5,4 @@
  *
  * @param health WallAnt 的初始生命值，默认为 4
  */
-WallAnt::WallAnt(int health) : Ant(health) {}
+WallAnt::WallAnt(double health) : Ant(health) {}
